+++ src/intel/common/gen_l3_config.c
@@ -245,7 +245,7 @@ gen_diff_l3_weights(struct gen_l3_weights w0, struct gen_l3_weights w1)
       float dw = 0;
 
       for (unsigned i = 0; i < GEN_NUM_L3P; i++)
-         dw += fabs(w0.w[i] - w1.w[i]);
+         dw += fabsf(w0.w[i] - w1.w[i]);
 
       return dw;
    }
